remove compile warning and don't create :ok tuple everytime
lib/tesla/adapter/ibrowse.ex:48: warning: the underscored variable
"_body" is used after being set. A leading underscore indicates that
the value of the variable should be ignored. If this is intended please
rename the variable to remove the underscore
lib/tesla/adapter/ibrowse.ex:51: warning: the underscored variable
"_body" is used after being set. A leading underscore indicates that
the value of the variable should be ignored. If this is intended please
rename the variable to remove the underscore
lib/tesla/adapter/ibrowse.ex:52: warning: the underscored variable
"_body" is used after being set. A leading underscore indicates that
the value of the variable should be ignored. If this is intended please
rename the variable to remove the underscore
lib/tesla/adapter/ibrowse.ex:57: warning: the underscored variable
"_status" is used after being set. A leading underscore indicates that
the value of the variable should be ignored. If this is intended please
rename the variable to remove the underscore
lib/tesla/adapter/ibrowse.ex:57: warning: the underscored variable
"_headers" is used after being set. A leading underscore indicates that
the value of the variable should be ignored. If this is intended please
rename the variable to remove the underscore