Start the :ssl application
Adds the `:ssl` application to the list of applications in `mix.exs`.

When using the default `httpc` adapter, attempting to GET an https URL would result in the error `** (Tesla.Error) adapter error: :econnrefused`.

The `:econnrefused` error is actually hiding the root cause, which is `:ssl_not_started`.

It appears that the httpc adapter is not automatically starting the `:ssl` application.