Allows put_headers to work with a nil header argument
I don't know if this is the best way to fix it or not, but I thought I
would put something out there to get the conversation going. I don't
like having `nil` passed in place of an empty list, but that was already
happening above.

Amos King @adkron <amos@binarynoggin.com>