Include `Tesla.Adapter.Mint` timeout's default value in documentation
I thought this might be helpful for someone else in the future. I spent some time looking for the default value for `timeout` in Hexdocs until I found it in the source code.
I've also mentioned that the value is in `milliseconds` to be more explicit.

Let me know what you think. Thanks!