Allow Tesla.Middleware.Fuse options to set the ask mode
Under high load, forcing all calls to `:fuse.ask/2` to be synchronous (sent
through the `fuse_server` process) can result in a deep mailbox and timeouts
inside the middleware (hitting the default call timeout of 5s). This commit adds
a `mode` option to the middleware that lets the user select `:async_dirty` so
that the `fuse_server` process can be bypassed.

In the documentation, I tried to be explicit that choosing `:async_dirty` might
result in drift, meaning that a given process might read the fuse state from
`ets` and it would not account for melts or resets that are in the mailbox of
the `fuse_server` process.

Because of potential side-effect/timing issues, it wasn't clear how to test this
change. Any guidance there is appreciated!