Add `store_opts` to allow multi-tenant caches
Implementation is a bit hacky, passing forward {store, opts} tuples
around, but enought to work and see how the API looks like.

Not entirely happey with `Store` behaviour (it's getting too wide), but
not much that can be done. One potential is to work via some sort of
Registry, but that may overcomplicate the middleware as impose some
avoidable overhead.