fix: keep original content length and encoding headers for HEAD requests
HEAD requests can be used to check the size of remote content
to decide ahead of time whether it is worth fetching. Of course
the size after decompression likely differs from the transfer size
indicated in the content-length header, but depending on use case
only the transfer size might be relevant.

This obsoletes the empty-body special case in decompress_body
previously added in 5bc9b82823b3238257619ea3d67f0985a3707d2b
since HEAD requests are now handled earlier. If we get an
invalid empty body in a non-HEAD request we want to fail loudly.